LayoutTests/media/track/track-webvtt-tc002-bom.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>
            
            function trackLoaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "2");
                
                testExpected("video.textTracks[0].cues[0].id","1");
                testExpected("video.textTracks[0].cues[0].startTime","0.0");
                testExpected("video.textTracks[0].cues[0].endTime","30.5");
                testExpected("video.textTracks[0].cues[0].getCueAsSource()", "Bear is Coming!!!!!");
                
                testExpected("video.textTracks[0].cues[1].id","2");
                testExpected("video.textTracks[0].cues[1].startTime","31.0");
                testExpected("video.textTracks[0].cues[1].endTime","1200.5");
                testExpected("video.textTracks[0].cues[1].getCueAsSource()", "I said Bear is coming!!!!");

                endTest();
            }
        </script>
    </head>
    <body>
        <p>Tests that the parser properly ignores a UTF-8 BOM character at the beginning of a file and all other cues are properly parsed.</p>
        <video>
            <track src="captions/tc002_bom.vtt" onload="trackLoaded()">
        </video>
    </body>
</html>
