LayoutTests/media/track/track-webvtt-tc002-bom-expected.txt

Tests that the parser properly ignores a UTF-8 BOM character at the beginning of a file and all other cues are properly parsed.

EXPECTED (video.textTracks[0].cues.length == '2') OK
EXPECTED (video.textTracks[0].cues[0].id == '1') OK
EXPECTED (video.textTracks[0].cues[0].startTime == '0.0') OK
EXPECTED (video.textTracks[0].cues[0].endTime == '30.5') OK
EXPECTED (video.textTracks[0].cues[0].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (video.textTracks[0].cues[1].id == '2') OK
EXPECTED (video.textTracks[0].cues[1].startTime == '31.0') OK
EXPECTED (video.textTracks[0].cues[1].endTime == '1200.5') OK
EXPECTED (video.textTracks[0].cues[1].getCueAsSource() == 'I said Bear is coming!!!!') OK
END OF TEST
